Move all new functionality into its own module
Not sure how this will all work once I need to make use of that render()
function in fx.js:start()
